+++ changelog.md
@@ -1,3 +1,11 @@
+##### 1.1.0
+- Added Reach Control, allowing interaction distance limits for Doors, Stairways, and Tokens
+- With Reach Control, the module is no longer limited to PF2e
+- Features that depend on a specific system or module can only be enabled when all requirements are met
+- Improved the Settings interface
+
+Reach Control was originally a fork of the Arms Reach module by p4535992. The version included in this module has been partially rewritten
+
 ##### 1.0.1 - 1.0.3
 - Small bug fixes
 - Some changes that are not user-related